remove word count as (potential) control variable
After some consideration, I have decided to not consider word count as
control variable. I think there is good reason to think that
repetition might be _causing_ low word count, due to the way story
excerpts for the study were prepared:
The raw story outputs produced by the model was cleaned by removing
unfinished sentences. However, once the model gets into a habit of
repetition it likes to form lists of repeating words, which can be quite
long-running. As a result, it is quite likely that shorter word length
was caused by the removal of a long, unfinished sentence, with plenty of
repetition. In this case, the relationship would be:
Y->Z (repetition->word count)
which is _not_ what you want for a control variable. Controlling for
these sort of variables can lead to spurious results (for an
explanation see Cinelli, Forney & Pearl, 2022: A Crash Course in Good
and Bad Controls).

Hence, controlling for word count could potentially do more harm than
good. I therefore dropped it from consideration.